services/backend/requirements.txt
fastapi
aiosqlite==0.19.0
pydantic==2.6.0
databases==0.8.0
uvicorn
passlib==1.7.4
python-multipart
python-jose
bcrypt==4.0.1
pyjwt==2.8.0
SQLAlchemy==1.4.51

